refactor: share document graph utilities
